requirements.txt
fastapi==0.133.0
lightgbm==4.6.0
twilio==9.10.2
python-multipart==0.0.22
httpx==0.28.1
slowapi==0.1.9
uvicorn==0.41.0
scikit-learn==1.8.0
joblib==1.5.3
pandas==3.0.1
numpy==2.4.2
pydantic==2.12.5
tldextract==5.3.1
requests==2.32.5
langdetect==1.0.9
python-telegram-bot==21.0.1
tqdm==4.67.3
pytesseract==0.3.13
Pillow==11.2.1
pytest==9.0.2
coverage==7.13.4
